Ehi:
Your code is thorough and looks very interactive.  Very professional-looking and quality work you did.

I was wondering: What led to your decision to use the lower case "i" added to 1 within the curly brace within the print command and to incorporate it into "for i in range..." on line 64? Does the i symbolize a function? I notice it matches the sky blue color coded components.  I think of the complex numbers I learned in algebra 2 when I see the "i", but I'm just curious here.